kisankiawaz-backend/requirements-base.txt
# requirements-base.txt
# Shared across ALL 12 services — installed once in kisan-base:latest

fastapi>=0.115.0
uvicorn[standard]>=0.34.0
firebase-admin>=6.6.0
google-cloud-firestore>=2.19.0
redis>=5.2.1
httpx>=0.28.0
bcrypt>=4.2.0
PyJWT>=2.10.0
pydantic[email]>=2.7.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
loguru>=0.7.3
